sent page and logic

+++ src/TronWallet.Core/Interfaces/Services/ITronGridClient.cs
@@ -5,7 +5,7 @@ namespace TronWallet.Core.Interfaces.Services;
 public interface ITronGridClient
 {
     Task<TronAccountResponse?>          GetAccountAsync(string base58Address);
-    Task<TronUnsignedTx>                CreateTransactionAsync(string fromHex, string toHex, long amountSun);
+    Task<TronUnsignedTx>                CreateTransactionAsync(string fromHex, string toHex, decimal amountSun);
     Task<TronBroadcastResponse>         BroadcastTransactionAsync(object signedTx);
     Task<TronTransactionInfo?>          GetTransactionInfoAsync(string txHashHex);
     // Task<TronTransactionsListResponse>  GetTransactionsAsync(string base58Address, int limit = 20, string? fingerprint = null);
